Consolidate NCEP GRIB tests and finalize Aero modernization
- Refactored `NCEPGribReader` to follow the Aero Protocol (lazy coordinates, scientific hygiene).
- Consolidated unit tests into `tests/test_aero_grib2.py` as requested.
- Removed redundant `tests/test_aero_ncep_grib.py`.
- Fixed CI linting issues and addressed reviewer feedback on robust coordinate renaming.
